requirements.txt
psycopg2-binary
dj-database-url
python-decouple
djangorestframework
drf-yasg
Django==4.0.2
django-cacheops==6.2
django-cleanup==7.0.0
django-cors-headers==4.4.0
django-crum==0.7.9
django-encrypted-model-fields==0.6.5
django-environ==0.10.0
django-filter==23.2
django-prometheus==2.3.1
django-redis==5.4.0
django-tenants==3.6.1
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
django-extensions
djangorestframework-simplejwt
requests